Terrain handling for 1.17/1.18 & Bugs fix for “ClickWindow“ packet

+++ MinecraftClient/ChatBot.cs
@@ -327,7 +327,7 @@ namespace MinecraftClient
         /// <param name="action">0 to create/update an item. 1 to remove an item.</param>
         /// <param name="objectivename">The name of the objective the score belongs to</param>
         /// <param name="value">The score to be displayed next to the entry. Only sent when Action does not equal 1.</param>
-        public virtual void OnUpdateScore(string entityname, byte action, string objectivename, int value) { }
+        public virtual void OnUpdateScore(string entityname, int action, string objectivename, int value) { }
 
         /// <summary>
         /// Called when an inventory/container was updated by server
